+++ Assets/Scripts/World/Movement/IWorldGridMap.cs
@@ -7,5 +7,6 @@ namespace Monsters.World.Movement
         bool IsInBounds(Vector2Int position);
         bool IsWalkable(Vector2Int position);
         WorldTileType GetTileType(Vector2Int position);
+        InteractionType GetInteractionType(Vector2Int position);
     }
 }